Add new functionality in the Chapter 2

+++ Chapter_02/SimpleMvcApplication/SimpleMvcApplication.Tests/Helpers/NumHelperTests.cs
@@ -8,13 +8,20 @@ namespace SimpleMvcApplication.Tests.Helpers
     {
         private readonly NumHelper _helper = new NumHelper();
 
+        /// <summary>
+        /// Test method - Factorial
+        /// Factorial method returns results then equals product of all positive integers less than or equal to given number.
+        /// Input parameter positive number 5.
+        /// Expected result 120.
+        /// Positive test.
+        /// </summary>
         [TestMethod]
         public void GetFactorialOfPositiveNumberCheckResultPositiveTest()
         {
             const int number = 5;
-            const int expectedResult = 120;
+            const long expectedResult = 120;
 
-            int actualResult = _helper.Factorial(number);
+            long actualResult = _helper.Factorial(number);
 
             Assert.AreEqual(expectedResult, actualResult);
         }
